backend-py/app/requirements.txt
fastapi
uvicorn[standard]
supabase
strawberry-graphql
pandas
numpy
scipy
requests
prisma
pyautogen
openai
langchain == 0.2.6
langchain-community == 0.2.6
langchain-anthropic == 0.1.19
langchain-openai == 0.1.14
langchainhub
